(FE) [refactor] dayjs 패키지 설치 밒 formatDate 분리.
- ReviewBox에서 formatDate 함수 분리하여 format.ts로 옮김.
- dayjs 패키지 설치하여 적용.
