+++ client/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.7",
     "axios": "^1.7.7",
+    "dayjs": "^1.11.13",
     "dotenv": "^16.4.5",
     "lottie-react": "^2.4.0",
     "npm-run-all": "^4.1.5",
